+++ packages/in_app_purchase/lib/src/in_app_purchase_connection/google_play_connection.dart
@@ -131,8 +131,7 @@ class GooglePlayConnection
         responses.expand((PurchasesResultWrapper response) {
       return response.purchasesList;
     }).map((PurchaseWrapper purchaseWrapper) {
-      return purchaseWrapper.toPurchaseDetails(
-          originalPurchaseID: purchaseWrapper.orderId);
+      return purchaseWrapper.toPurchaseDetails();
     }).toList();
 
     return QueryPurchaseDetailsResponse(
@@ -146,13 +145,10 @@ class GooglePlayConnection
     );
   }
 
-  /// This is a non-op.
-  ///
-  /// There is no refreshing verification data on Google Play.
   @override
-  Future<PurchaseVerificationData> refreshPurchaseVerificationData(
-      PurchaseDetails purchase) async {
-    return purchase.verificationData;
+  Future<PurchaseVerificationData> refreshPurchaseVerificationData() async {
+    throw Exception(
+        'The method <refreshPurchaseVerificationData> only works on iOS.');
   }
 
   @override